refactor: use own proto package
