+++ apps/admin/package.json
@@ -14,13 +14,12 @@
 		"@energyleaf/db": "workspace:^",
 		"@energyleaf/lib": "workspace:^",
 		"@energyleaf/mail": "workspace:*",
-		"@energyleaf/proto": "workspace:*",
+		"@energyleaf/proto": "^0.1.6",
 		"@energyleaf/tailwindcss": "workspace:*",
 		"@energyleaf/ui": "workspace:*",
 		"@hookform/resolvers": "^3.4.0",
 		"@node-rs/argon2": "^1.8.3",
 		"@node-rs/bcrypt": "^1.10.4",
-		"@protobuf-ts/runtime": "^2.9.4",
 		"@t3-oss/env-core": "^0.10.1",
 		"@t3-oss/env-nextjs": "^0.10.1",
 		"@tanstack/react-table": "^8.17.3",
