--- packages/proto/protos/energyleaf.options
@@ -1,6 +1,0 @@
-TokenRequest.client_id max_size:128
-TokenResponse.access_token max_size:45
-TokenResponse.status_message max_size:255
-TokenResponse.script max_size:255
-SensorDataRequest.access_token max_size:45
-SensorDataResponse.status_message max_size:255
