--- packages/proto/tsconfig.json
@@ -1,3 +1,0 @@
-{
-	"extends": "@energyleaf/tsconfig/base.json"
-}
